listo la categorizacion cargada por ws se corrigio tambien el costo double funciona eeeeeeeeeeeeee XD

+++ app/src/main/java/com/ipxserver/davidtorrez/fvposv0/Listeners/FragmentReceiver.java
@@ -24,7 +24,7 @@ public class FragmentReceiver extends BroadcastReceiver
     private final GridbarAdapter gridAdapter;
     private final PrincipalActivity main;
     private ArrayList<Product> listaProductos=null;
-    private int monto;
+    private Double monto;
     public FragmentReceiver( GridbarAdapter gridAdapter,PrincipalActivity main)
     {
         this.gridAdapter = gridAdapter;
@@ -42,7 +42,7 @@ public class FragmentReceiver extends BroadcastReceiver
         if(fragment_id==FRAGMENT_FACTURA)
         {
             listaProductos = (ArrayList<Product>)intent.getSerializableExtra("lista_seleccionados");
-            monto = intent.getIntExtra("monto",-1);
+            monto = intent.getDoubleExtra("monto",-1);
             Log.i("David", "tamao del list product size " + listaProductos.size());
 //            Product product = new Product();
 //            product.setCost("23 ");
@@ -63,7 +63,7 @@ public class FragmentReceiver extends BroadcastReceiver
         return listaProductos;
     }
 
-    public int getMonto() {
+    public Double getMonto() {
         return monto;
     }
 }
